+++ libs/sdk/testing/projects/testing/src/a11y/a11y-analyzer.ts
@@ -69,7 +69,7 @@ export abstract class SkyA11yAnalyzer {
       };
 
       SkyA11yAnalyzer.analyzer.run(
-        element,
+        element!,
         {...defaults, ...config},
         callback
       );