+++ libs/sdk/testing/projects/testing/src/matchers/matchers.spec.ts
@@ -1,7 +1,10 @@
+import { HttpClientTestingModule } from '@angular/common/http/testing';
 import {
-  async,
+  fakeAsync,
   inject,
-  TestBed
+  TestBed,
+  tick,
+  waitForAsync
 } from '@angular/core/testing';
 
 import {
@@ -60,7 +63,7 @@ describe('Jasmine matchers', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [SkyI18nModule],
+      imports: [HttpClientTestingModule, SkyI18nModule],
       providers: [SkyAppResourcesService]
     });
     document.body.innerHTML = '';
@@ -192,12 +195,14 @@ describe('Jasmine matchers', () => {
 
   describe('toBeAccessible', () => {
 
-    it('should check accessibility', async(() => {
+    it('should check accessibility', waitForAsync(() => {
       const element = createPassingElement();
-      expect(element).toBeAccessible();
+      expect(element).toBeAccessible(() => {
+        expect(element).toExist();
+      });
     }));
 
-    it('should fail if accessibility rules fail', async(() => {
+    it('should fail if accessibility rules fail', waitForAsync(() => {
       const failSpy = spyOn((window as any), 'fail').and.callFake((message: string) => {
         expect(message.indexOf('duplicate-id') > -1).toEqual(true);
       });
@@ -228,19 +233,19 @@ describe('Jasmine matchers', () => {
         });
       });
 
-      it('should allow configuration override', async(() => {
+      it('should allow configuration override', waitForAsync(() => {
         const element = createFailingElement();
-        expect(element).toBeAccessible(() => { }, {
+        expect(element).toBeAccessible(() => {}, {
           rules: {
             'duplicate-id': { enabled: false }
           }
         });
       }));
 
-      it('should allow SkyAppConfig override', async(
+      it('should allow SkyAppConfig override', waitForAsync(
         inject([SkyAppConfig], (config: SkyAppConfig) => {
           const element = createPassingElement();
-          expect(element).toBeAccessible(() => { }, config.skyux.a11y as SkyA11yAnalyzerConfig);
+          expect(element).toBeAccessible(() => {}, config.skyux.a11y as SkyA11yAnalyzerConfig);
         }))
       );
     });
@@ -253,7 +258,7 @@ describe('Jasmine matchers', () => {
       resourcesService = TestBed.inject(SkyAppResourcesService);
     });
 
-    it('should check that the actual text matches text provided by resources', async(() => {
+    it('should check that the actual text matches text provided by resources', waitForAsync(() => {
       const messageKey = 'name';
       const messageValue = 'message from resource';
       const text = 'message from resource';
@@ -269,7 +274,7 @@ describe('Jasmine matchers', () => {
       expect(text).toEqualResourceText(messageKey);
     }));
 
-    it('should check that the actual text matches text provided by resources with arguments', async(() => {
+    it('should check that the actual text matches text provided by resources with arguments', waitForAsync(() => {
       const messageKey = 'nameWithArgs';
       const messageValue = 'message from resources with args = {0}';
       const messageArgs: any[] = [100];
@@ -283,7 +288,7 @@ describe('Jasmine matchers', () => {
         }
       });
 
-      expect(text).toEqualResourceText(messageKey, messageArgs, () => { });
+      expect(text).toEqualResourceText(messageKey, messageArgs, () => {});
     }));
 
     it('should fail if the actual text does not match text provided by resources', (done) => {
@@ -314,7 +319,7 @@ describe('Jasmine matchers', () => {
       resourcesService = TestBed.inject(SkyAppResourcesService);
     });
 
-    it('should check that the element\'s text matches text provided by resources', async(() => {
+    it('should check that the element\'s text matches text provided by resources', waitForAsync(() => {
       const messageKey = 'name';
       const messageValue: string = 'message from resource';
       const elem = createElement(messageValue);
@@ -330,7 +335,7 @@ describe('Jasmine matchers', () => {
       expect(elem).toHaveResourceText(messageKey);
     }));
 
-    it('should default to trimming whitespace and check that the element\'s text matches text provided by resources', async(() => {
+    it('should default to trimming whitespace and check that the element\'s text matches text provided by resources', async () => {
       const messageKey = 'name';
       const messageValue: string = 'message from resource';
       const elem: any = createElement(`    ${messageValue}     `);
@@ -344,9 +349,9 @@ describe('Jasmine matchers', () => {
       });
 
       expect(elem).toHaveResourceText(messageKey);
-    }));
+    });
 
-    it('should check that the element\'s text matches text provided by resources with arguments', async(() => {
+    it('should check that the element\'s text matches text provided by resources with arguments', async () => {
       const messageKey = 'nameWithArgs';
       const messageValue = 'message from resources with args = {0}';
       const messageArgs: any[] = [100];
@@ -361,7 +366,7 @@ describe('Jasmine matchers', () => {
       });
 
       expect(elem).toHaveResourceText(messageKey, messageArgs);
-    }));
+    });
 
     it('should fail if the element\'s text does not match text provided by resources', (done) => {
       const messageKey = 'nameThatDoesNotExist';